Create smartsplits.sol
This smart contract has the following functionality:
Ownership: The contract is owned by the msg.sender who deploys the contract and the owner can authorize or revoke other users.
Wallet addresses: The owner can set the two wallet addresses using the setWallet1 and setWallet2 functions. These addresses are stored in the contract and can be changed at any time by the owner.
Percentage allocation: The owner can set the percentage allocation using the setPercentage function. This percentage is used to determine how much of the received funds should be sent to the first wallet address and how much should be sent to the second wallet address.
Fund receipt: When the contract receives Ether, it automatically distributes the funds to the two wallet addresses according to the percentage allocation set by the owner. The contract will check if the wallet addresses and percentage allocation are set before receiving funds, otherwise, it will reject the transaction.
Access control: Only authorized users can set the wallet addresses and percentage allocation, and the owner can add or remove authorized users.
Event logging: The contract emits events when the wallet addresses and percentage allocation are set, and when funds are received and allocated to the two wallets. This allows for tracking of the transactions and to check which addresses were authorized, who set the address and percentage and also check the amount of funds received and transferred.
Security measures: The contract has several security measures such as checking that the msg.sender is authorized before allowing the wallet addresses and percentage allocation to be set, and that the wallet addresses and percentage allocation are set before receiving funds, also checking that the percentage is less than or equal to 100, and more.